BLP {hdm}	R Documentation

BLP data set

Description

Automobile data set from the US.

Format

model.name

model name

model.id

model id

firm.id

firm id

cdid

cdid

id

id

price

log price

mpg

miles per gallon

mpd

miles per dollar

hpwt

horse power per weight

air

air conditioning (binary variable)

space

size of the car

share

market share

outshr

share s0

y

outcome variable defined as log(share) - log(outshr)

trend

time trend

Details

Data set was analysed in Berry, Levinsohn and Pakes (1995). The data stem from annual issues of the Automotive News Market Data Book. The data set inlcudes information on all models marketed during the the periord beginning 1971 and ending in 1990 cotaining 2217 model/years from 997 distinct models. A detailed description is given in BLP (1995, 868–871). The internal function constructIV constructs instrumental variables along the lines described and used in BLP (1995).

References

S. Berry, J. Levinsohn, A. Pakes (1995). Automobile Prices in Market EquilibriumD. Econometrica, 63(4), 841–890.

Examples

data(BLP)

[Package hdm version 0.2.0 Index]