Growth Data {hdm}	R Documentation

Growth data set

Description

Data set of growth compiled by Barro Lee.

Format

Dataframe with the following variables:

outcome

dependent variable: national growth rates in GDP per capita for the periods 1965-1975 and 1975-1985

x

covariates which might influence growth

Details

The data set contains growth data of Barro-Lee. The Barro Lee data consists of a panel of 138 countries for the period 1960 to 1985. The dependent variable is national growth rates in GDP per capita for the periods 1965-1975 and 1975-1985. The growth rate in GDP over a period from t_1 to t_2 is commonly defined as \log(GDP_{t_1}/GDP_{t_2}). The number of covariates is p=62. The number of complete observations is 90.

Source

The full data set and further details can be found at http://www.nber.org/pub/barro.lee, http://www.barrolee.com, and, http://www.bristol.ac.uk//Depts//Economics//Growth//barlee.htm.

References

R.J. Barro, J.W. Lee (1994). Data set for a panel of 139 countries. NBER.

R.J. Barro, X. Sala-i-Martin (1995). Economic Growth. McGrwa-Hill, New York.

Examples

data(GrwothData)

[Package hdm version 0.2.0 Index]