LassoShooting.fit {hdm}	R Documentation

Shooting Lasso

Description

Implementation of the Shooting Lasso (Fu, 1998) with variable dependent penalization weights.

Usage

LassoShooting.fit(x, y, lambda, control = list(maxIter = 1000, optTol =
  10^(-5), zeroThreshold = 10^(-6)), XX = NULL, Xy = NULL,
  beta.start = NULL)

Arguments

x
matrix of regressor variables (n times p where n denotes the number of observations and p the number of regressors)

y
dependent variable (vector or matrix)

lambda
vector of length p of penalization parameters for each regressor

control
list with control parameters: maxIter maximal number of iterations, optTol tolerance for parameter precision, zeroThreshold threshold applied to the estimated coefficients for numerical issues.

XX
optional, precalculated matrix t(X)*X

Xy
optional, precalculated matrix t(X)*y

beta.start
start value for beta

Details

The function implements the Shooting Lasso (Fu, 1998) with variable dependent penalization. The arguments XX and Xy are optional and allow to use precalculated matrices which might improve performance.

Value

coefficients
estimated coefficients by the Shooting Lasso Algorithm

coef.list
matrix of coefficients from each iteration

num.it
number of iterations run

References

Fu, W. (1998). Penalized regressions: the bridge vs the lasso. Journal of Computational and Graphical Software 7, 397-416.

[Package hdm version 0.2.0 Index]