rlassoATE {hdm}	R Documentation

Functions for estimation of treatment effects

Description

This class of functions estimates the average treatment effect (ATE), the ATE of the tretated (ATET), the local average treatment effects (LATE) and the LATE of the tretated (LATET). The estimation methods rely on immunized / orthogonal moment conditions which guarantee valid post-selection inference in a high-dimensional setting. Further details can be found in Belloni et al. (2014).

Usage

rlassoATE(x, ...)

## Default S3 method:
rlassoATE(x, d, y, bootstrap = "none", nRep = 500, ...)

## S3 method for class 'formula'
rlassoATE(formula, data, bootstrap = "none", nRep = 500,
  ...)

rlassoATET(x, ...)

## Default S3 method:
rlassoATET(x, d, y, bootstrap = "none", nRep = 500, ...)

## S3 method for class 'formula'
rlassoATET(formula, data, bootstrap = "none", nRep = 500,
  ...)

rlassoLATE(x, ...)

## Default S3 method:
rlassoLATE(x, d, y, z, bootstrap = "none", nRep = 500,
  post = TRUE, intercept = TRUE, ...)

## S3 method for class 'formula'
rlassoLATE(formula, data, bootstrap = "none", nRep = 500,
  post = TRUE, intercept = TRUE, ...)

rlassoLATET(x, ...)

## Default S3 method:
rlassoLATET(x, d, y, z, bootstrap = "none", nRep = 500,
  post = TRUE, intercept = TRUE, ...)

## S3 method for class 'formula'
rlassoLATET(formula, data, bootstrap = "none", nRep = 500,
  post = TRUE, intercept = TRUE, ...)

Arguments

x
exogenous variables

...
arguments passed, e.g. intercept and post

d
treatment variable (binary)

y
outcome variable / dependent variable

bootstrap
boostrap method which should be employed: 'none', 'Bayes', 'normal', 'wild'

nRep
number of replications for the bootstrap

formula
An object of class Formula of the form " y ~ x + d | x" with y the outcome variable, d treatment variable, and x exogenous variables.

data
An optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model. If not found in data, the variables are taken from environment(formula), typically the environment from which rlassoATE is called.

z
instrumental variables (binary)

post
logical. If TRUE, post-lasso estimation is conducted.

intercept
logical. If TRUE, intercept is included which is not penalized.

Details

Details can be found in Belloni et al. (2014).

Value

Functions return an object of class rlassoTE with estimated effects, standard errors and individual effects in the form of a list.

References

A. Belloni, V. Chernozhukov, I. Fernandez-Val, and C. Hansen (2014). Program evaluation with high-dimensional data. Working Paper.

[Package hdm version 0.2.0 Index]