cps2012 {hdm}	R Documentation

cps2012 data set

Description

Census data from the US for the year 2012.

Format

lnw

log of hourly wage (annual earnings / annual hours)

female

female indicator

married status

six indicators: widowed, divorced, separated, nevermarried, and married (omitted)

education attainment

six indicators: hsd08, hsd911, hsg, cg, ad, and sc (omitted)

region indicators

four indicators: mw, so, we, and ne (omitted)

potential experience

(max[0, age - years of education - 7]): exp1, exp2 (divided by 100), exp3 (divided by 1000), exp4 (divided by 10000)

weight

March Supplement sampling weight

year

CPS year

Details

The CPS is a monthly U.S. household survey conducted jointly by the U.S. Census Bureau and the Bureau of Labor Statistics. The data comprise the year 2012. This data set was used in Mulligan and Rubinstein (2008). The sample comprises white non-hipanic, ages 25-54, working full time full year (35+ hours per week at least 50 weeks), exclude living in group quarters, self-employed, military, agricultural, and private household sector, allocated earning, inconsistent report on earnings and employment, missing data.

References

C. B. Mulligan and Y. Rubinstein (2008). Selection, investment, and women's relative wages over time. The Quarterly Journal of Economics, 1061–1110.

Examples

data(BLP)

[Package hdm version 0.2.0 Index]