lambdaCalculation {hdm}	R Documentation

Function for Calculation of the penalty parameter

Description

This function implements different methods for calculation of the penalization parameter λ. Further details can be found under rlasso.

Usage

lambdaCalculation(penalty = list(homoscedastic = FALSE, X.dependent.lambda =
  FALSE, lambda.start = NULL, c = 1.1, gamma = 0.1), y = NULL, x = NULL)

Arguments

penalty
list with options for the calculation of the penalty.

c and gamma constants for the penalty with default c=1.1 and gamma=0.1

homoscedastic logical, if homoscedastic errors are considered (default FALSE). Option none is described below.

X.dependent.lambda if independent or dependent design matrix X is assumed for calculation of the parameter λ

numSim number of simulations for the X-dependent methods

lambda.start initial penalization value, compulsory for method "none"

y
residual which is used for calculation of the variance or the data-dependent loadings

x
matrix of regressor variables

Value

The functions returns a list with the penalty lambda which is the product of lambda0 and Ups0. Ups0 denotes either the variance (independent case) or the data-dependent loadings for the regressors. method gives the selected method for the calculation.

[Package hdm version 0.2.0 Index]