print.rlassoEffects {hdm}	R Documentation

Methods for S3 object rlassoEffects

Description

Objects of class rlassoEffects are constructed by rlassoEffects. print.rlassoEffects prints and displays some information about fitted rlassoEffect objects. summary.rlassoEffects summarizes information of a fitted rlassoEffect object and is described at summary.rlassoEffects. confint.rlassoEffects extracts the confidence intervals. plot.rlassoEffects plots the estimates with confidence intervals.

Usage

## S3 method for class 'rlassoEffects'
print(x, digits = max(3L, getOption("digits") - 3L),
  ...)

## S3 method for class 'rlassoEffects'
confint(object, parm, level = 0.95, joint = FALSE,
  ...)

## S3 method for class 'rlassoEffects'
plot(x, joint = FALSE, level = 0.95, main = "",
  xlab = "coef", ylab = "", xlim = NULL, ...)

Arguments

x
an object of class rlassoEffects

digits
significant digits in printout

...
arguments passed to the print function and other methods.

object
an object of class rlassoEffects

parm
a specification of which parameters are to be given confidence intervals among the variables for which inference was done, either a vector of numbers or a vector of names. If missing, all parameters are considered.

level
confidence level required

joint
logical, if TRUE joint confidence intervals are calculated.

main
an overall title for the plot

xlab
a title for the x axis

ylab
a title for the y axis

xlim
vector of length two giving lower and upper bound of x axis

[Package hdm version 0.2.0 Index]