print.rlassoIV {hdm}	R Documentation

Methods for S3 object rlassoIV

Description

Objects of class rlassoIV are constructed by rlassoIV. print.rlassoIV prints and displays some information about fitted rlassoIV objects. summary.rlassoIV summarizes information of a fitted rlassoIV object. confint.rlassoIV extracts the confidence intervals.

Usage

## S3 method for class 'rlassoIV'
print(x, digits = max(3L, getOption("digits") - 3L), ...)

## S3 method for class 'rlassoIV'
summary(object, digits = max(3L, getOption("digits") - 3L),
  ...)

## S3 method for class 'rlassoIV'
confint(object, parm, level = 0.95, ...)

Arguments

x
an object of class rlassoIV

digits
significant digits in printout

...
arguments passed to the print function and other methods

object
An object of class rlassoIV

parm
a specification of which parameters are to be given confidence intervals, either a vector of numbers or a vector of names. If missing, all parameters are considered.

level
confidence level required.

[Package hdm version 0.2.0 Index]