print.rlassoTE {hdm}	R Documentation

Methods for S3 object rlassoTE

Description

Objects of class rlassoTE are constructed by rlassoATE, rlassoATET, rlassoLATE, rlassoLATET. print.rlassoTE prints and displays some information about fitted rlassoTE objects. summary.rlassoTE summarizes information of a fitted rlassoTE object. confint.rlassoTE extracts the confidence intervals.

Usage

## S3 method for class 'rlassoTE'
print(x, digits = max(3L, getOption("digits") - 3L), ...)

## S3 method for class 'rlassoTE'
summary(object, digits = max(3L, getOption("digits") - 3L),
  ...)

## S3 method for class 'rlassoTE'
confint(object, parm, level = 0.95, ...)

Arguments

x
an object of class rlassoTE

digits
number of significant digits in printout

...
arguments passed to the print function and other methods

object
an object of class rlassoTE

parm
a specification of which parameters are to be given confidence intervals, either a vector of numbers or a vector of names. If missing, all parameters are considered.

level
confidence level required.

[Package hdm version 0.2.0 Index]