predict.rlassologit {hdm}	R Documentation

Methods for S3 object rlassologit

Description

Objects of class rlassologit are constructed by rlassologit. print.rlassologit prints and displays some information about fitted rlassologit objects. summary.rlassologit summarizes information of a fitted rlassologit object. predict.rlassologit predicts values based on a rlassologit object. model.matrix.rlassologit constructs the model matrix of a lasso object.

Usage

## S3 method for class 'rlassologit'
predict(object, newdata = NULL, type = "response",
  ...)

## S3 method for class 'rlassologit'
model.matrix(object, ...)

## S3 method for class 'rlassologit'
print(x, all = TRUE, digits = max(3L,
  getOption("digits") - 3L), ...)

## S3 method for class 'rlassologit'
summary(object, all = TRUE, digits = max(3L,
  getOption("digits") - 3L), ...)

Arguments

object
an object of class rlassologit

newdata
new data set for prediction

type
type of prediction required. The default ('response) is on the scale of the response variable; the alternative 'link' is on the scale of the linear predictors.

...
arguments passed to the print function and other methods

x
an object of class rlassologit

all
logical, indicates if coefficients of all variables (TRUE) should be displayed or only the non-zero ones (FALSE)

digits
significant digits in printout

[Package hdm version 0.2.0 Index]