print.rlassologitEffects {hdm}	R Documentation

Methods for S3 object rlassologitEffects

Description

Objects of class rlassologitEffects are construced by rlassologitEffects or rlassologitEffect. print.rlassologitEffects prints and displays some information about fitted rlassologitEffect objects. summary.rlassologitEffects summarizes information of a fitted rlassologitEffects object. confint.rlassologitEffects extracts the confidence intervals.

Usage

## S3 method for class 'rlassologitEffects'
print(x, digits = max(3L, getOption("digits") -
  3L), ...)

## S3 method for class 'rlassologitEffects'
summary(object, digits = max(3L,
  getOption("digits") - 3L), ...)

## S3 method for class 'rlassologitEffects'
confint(object, parm, level = 0.95,
  joint = FALSE, ...)

Arguments

x
an object of class rlassologitEffects

digits
number of significant digits in printout

...
arguments passed to the print function and other methods

object
an object of class rlassologitEffects

parm
a specification of which parameters are to be given confidence intervals, either a vector of numbers or a vector of names. If missing, all parameters are considered.

level
confidence level required.

joint
logical, if joint confidence intervals should be clalculated

[Package hdm version 0.2.0 Index]