print.tsls {hdm}	R Documentation

Methods for S3 object tsls

Description

Objects of class tsls are constructed by tsls. print.tsls prints and displays some information about fitted tsls objects. summary.tsls summarizes information of a fitted tsls object.

Usage

## S3 method for class 'tsls'
print(x, digits = max(3L, getOption("digits") - 3L), ...)

## S3 method for class 'tsls'
summary(object, digits = max(3L, getOption("digits") - 3L),
  ...)

Arguments

x
an object of class tsls

digits
significant digits in printout

...
arguments passed to the print function and other methods

object
an object of class tsls

[Package hdm version 0.2.0 Index]