rlassologitEffects {hdm}	R Documentation

rigorous Lasso for Logistic Models: Inference

Description

The function estimates (low-dimensional) target coefficients in a high-dimensional logistic model.

Usage

rlassologitEffects(x, ...)

## Default S3 method:
rlassologitEffects(x, y, index = c(1:ncol(x)), I3 = NULL,
  post = TRUE, ...)

## S3 method for class 'formula'
rlassologitEffects(formula, data, I, included = NULL,
  post = TRUE, ...)

rlassologitEffect(x, y, d, I3 = NULL, post = TRUE)

Arguments

x
matrix of regressor variables serving as controls and potential treatments. For rlassologitEffect it contains only controls, for rlassologitEffects both controls and potential treatments. For rlassologitEffects it must have at least two columns.

y
outcome variable

index
vector of integers, logical or names indicating the position (column) or name of variables of x which should be used as treatment variables.

I3
logical vector with same length as the number of controls; indicates if variables (TRUE) should be included in any case.

post
logical. If TRUE, post-Lasso estimation is conducted.

formula
An element of class formula specifying the linear model.

data
an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model. If not found in data, the variables are taken from environment(formula), typically the environment from which the function is called.

I
An one-sided formula specifying the variables for which inference is conducted.

included
One-sided formula of variables which should be included in any case.

d
variable for which inference is conducted (treatment variable)

...
additional parameters

Details

The functions estimates (low-dimensional) target coefficients in a high-dimensional logistic model. An application is e.g. estimation of a treatment effect α_0 in a setting of high-dimensional controls. The function is a wrap function for rlassologitEffect which does inference for only one variable (d).

Value

The function returns an object of class rlassologitEffects with the following entries:

coefficients
estimated value of the coefficients

se
standard errors

t
t-statistics

pval
p-values

samplesize
sample size of the data set

I
index of variables of the union of the lasso regressions

References

A. Belloni, V. Chernozhukov, Y. Wei (2013). Honest confidence regions for a regression parameter in logistic regression with a loarge number of controls. cemmap working paper CWP67/13.

[Package hdm version 0.2.0 Index]