summary.rlassoEffects {hdm}	R Documentation

Summarizing rlassoEffects fits

Description

Summary method for class rlassoEffects

Usage

## S3 method for class 'rlassoEffects'
summary(object, ...)

## S3 method for class 'summary.rlassoEffects'
print(x, digits = max(3L, getOption("digits")
  - 3L), ...)

Arguments

object
an object of class rlassoEffects, usually a result of a call to rlassoEffects

...
further arguments passed to or from other methods.

x
an object of class summary.rlassoEffects, usually a result of a call or summary.rlassoEffects

digits
the number of significant digits to use when printing.

Details

Summary of objects of class rlassoEffects

[Package hdm version 0.2.0 Index]